+++ Braintree/API/Models/BTMutableApplePayPaymentMethod.h
@@ -1,6 +1,8 @@
+#if BT_ENABLE_APPLE_PAY
 #import "BTApplePayPaymentMethod.h"
 #import "BTApplePayPaymentMethod_Internal.h"
 
 @interface BTMutableApplePayPaymentMethod : BTApplePayPaymentMethod
 
 @end
+#endif