+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Braintree iOS SDK Release Notes
 
+## 3.6.0
+
+* Make Apple Pay a build option, enabled via `Braintree/Apple-Pay` subspec, 
+  which adds a `BT_ENABLE_APPLE_PAY=1` preprocesor macro.
+  * Addresses an issue [reported by developers attempting to submit v.zero integrations without Apple Pay to the app store](https://github.com/braintree/braintree_ios/issues/60).
+
 ## 3.5.0 (2014-12-03)
 
 * Add localizations to UI and Drop-In subspecs: